Makefile: optionally compile with clang
Different compilers can produce different warnings so having the option
to use gcc or clang lets us take advantage of that. Clang is also the
preferred compiler for Android projects.

The clang build fails at the moment but this adds the infrastructure to
select clang as the compiler with the `CLANG` environment variable:

    CLANG=1 make

Change-Id: I2dd5e8ce8e7fc20e93f5780348726a2f9b76239b