Make the clang build work.
The funky offset calculation was causing clang some problems. It
complained about the invalid assembly used to extract the offsets
calculated by the compiler.

This changes the funkiness from using a fake mneumonic (which clang
can't not complain about) to using a constant. The assembly post
processing is updated accordingly.

Change-Id: I5beef119e8f7cefedcf99642a28cf5aa5bbfdff8