Only allow identity memory mapping
The code has the assumption that virtual addresses are mapped to the
same physical address but the mm function would allow this assumption to
be broken. Only offering identity mapping function avoids this problem
by enforcing assumption.

Non-identity mapping can still be added later if there proves to be a
need for it.

Change-Id: Ie3db41e1582c72d880275acdb48cabefe05b3ba0