Move VM shared headers under vmapi/
This directory of headers is intended to be shared with the VMs and
includes the definitions needed for them to communicate.

The hypervisor includes them as e.g. `vmapi/hf/hvc.h` but the VMs will
have the `vmapi` directory in the include path so will just have to
include e.g. `hf/hvc.h`.

Change-Id: I32441bbaf20294c02ba5e71078e607f89f66e892