Run QEMU tests individually.
Each test should have a clean context to run in which means a fresh QEMU
run is needed. The test image will report the tests it contains and the
runner script will query this and then run the individual tests.

The test descriptors are stored in the .hftest section of the image to
allow the image to query itself.

Change-Id: I9d4d10097d768aa51728e428a4e48eaa9b4949b4