Decouple logs from seconday VM index.
As the primary VM gets given ID 0, the secondary index becomes off by
one. This could be patched where it is logged but it is also more
convenient for debugging if the name give to the VM is logged.

We don't have a way to log a memiter and the precision syntax is not
supported by dlog. We don't particularly want a fully fledged printf and
this is only used in one place at the moment so the print is done
manually. If we improve dlog in future, this can be updated.

Change-Id: I8192f9ab970438e150f82de0e4437ac3d37143cb