Remove single page mapping functions.
This was more code to maintain for no benefit. It was a specialized
version of the generic mapping and only used for mapping the PL011 so it
was also rarely run.

It's uses have been replaced with the generic mapping function but if we
can show that we need it later then it can be added back.

Change-Id: I836007dd690dcc9dbdc79bd424127139f20fc0a5