Support concatenated page tables.
To reduce the depth of the page table tree, some architectures
concatenate top level page tables rather than introducing a new level in
the tree.

The tests cover more of the memory management code and also document
some of the quirks that exist in the APIs. The tests also exercise the
concatenated tables logic.

Bug: 117549422
Change-Id: I99991aaf3bfb753dd6176cb9df7a5337ed9c184d