Bring back state update in api_switch_to_primary.
We need it to take vCPUs from running to ready state, otherwise vCPUs
cannot run anymore. It is safe to do so because we already require that
VM locks be acquired before vCPU locks (when they are held concurrently):
api_mailbox_send already does this.

Also added a test to catch errors when sending/receiving messages leaves
a vCPU is non-runnable state.

Change-Id: I18d9aed89224876abbb2df2039437bad43603ab1