Correct aarch64 PTE attribute handling.
Not all the non-address bits are attribute bits so the mask needs to
exclude those that are not. This was causing mapping to miss the case
when then entry is already in place as the attributes appeared
different.

Other cases continued to work by luck as the extra included bits did not
interfere with behaviour.

Change-Id: I8384e952cd964e245097d72909999ce6599c3736