Implement virtual timer for secondary VMs.
If the timer fires while the secondary VM is running, the physical
interrupt will go to the primary VM, which should then run the secondary
again. Hafnium will then inject the virtual timer interrupt
automatically, and mask the timer interrupt so it doesn't immediately go
back to the primary. Note that the secondary must clear this mask before
it can get another timer interrupt.

Bug: 117271574
Change-Id: If90a3c4fd133333713ad6bf7d6a3b515ae6fea6a