Remove MM_MODE_NOINVALIDATE.
This flag exists to avoid paying the cost required for consistency when
consistency is not required. It was used for stage-2 allocation when the
VMs are initially being created i.e. there is a point before which all
stage-2 updates don't invalidate and after which they do. Given this, it
can move to the same approach used by the locks for dlog and mpool
allowing the mode clutter to be removed from where is doesn't belong.

Change-Id: I9be9e42c1daaf8570d7555bb3f37876aa7c3bd51